Moving the implementation to the monadic way
